pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-orchestrator"
version = "1.0.0"
description = "CLI tool to orchestrate multiple AI models for automatic task distribution"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
]

[project.scripts]
ai-orchestrator = "ai_orchestrator.cli:main"

[tool.setuptools.packages.find]
where = ["."]